Division of tests into unit and integration
