+++ Makefile
@@ -10,6 +10,14 @@ test:			## Run tests.
 	make down
 	make build
 	docker-compose -f tests/docker-compose.yml run php-cli vendor/bin/phpunit
+test-unit:		## Run unit tests.
+	make down
+	make build
+	docker-compose -f tests/docker-compose.yml run php-cli vendor/bin/phpunit --testsuite Unit
+test-integration:	## Run integration tests.
+	make down
+	make build
+	docker-compose -f tests/docker-compose.yml run php-cli vendor/bin/phpunit --testsuite Integration
 static-analyze:		## Run code static analyze.
 	make build
 	docker-compose -f tests/docker-compose.yml run php-cli vendor/bin/psalm --config=psalm.xml --shepherd --stats --php-version=8.1
@@ -17,5 +25,6 @@ rector:			## Run rector.
 	make build
 	docker-compose -f tests/docker-compose.yml run php-cli vendor/bin/rector process --config=rector.php --dry-run
 mutation:		## Run mutation tests.
+	make down
 	make build
 	docker-compose -f tests/docker-compose.yml run php-cli ./vendor/bin/roave-infection-static-analysis-plugin
