+++ tests/Unit/QueueProviderTest.php
@@ -11,25 +11,14 @@ use Yiisoft\Queue\Redis\QueueProviderInterface;
 class QueueProviderTest extends TestCase
 {
 
-    /**
-     * @depends test__construct
-     */
-    public function testGetId(QueueProvider $provider)
-    {
-        $id = $provider->getId();
-        $this->assertGreaterThan(0, $id);
-    }
-
     public function test__construct()
     {
-        $redis = new \Redis();
-        $connected = $redis->connect('redis');
-        $this->assertTrue($connected);
+        $redis = $this->createMock(\Redis::class);
+        $redis->method('isConnected')->willReturn(true);
         $provider = new QueueProvider($redis, 'test');
         $this->assertInstanceOf(QueueProviderInterface::class, $provider);
         return $provider;
     }
-
     /**
      * @depends test__construct
      * @throws \PHPUnit\Framework\MockObject\Exception
